fix: refactor root command logic
