--- command/completion.go
@@ -1,31 +1,0 @@
-package command
-
-import (
-	"fmt"
-
-	"github.com/spf13/cobra"
-)
-
-func NewCommandCompletion() *cobra.Command {
-	return &cobra.Command{
-		Use:       "completion [bash|zsh|fish|powershell]",
-		Short:     "Generate shell completion scripts",
-		Long:      "Generate shell completion scripts for your shell.",
-		Args:      cobra.MatchAll(cobra.ExactArgs(1), cobra.OnlyValidArgs),
-		ValidArgs: []string{"bash", "zsh", "fish", "powershell"},
-		RunE: func(cmd *cobra.Command, args []string) error {
-			switch args[0] {
-			case "bash":
-				return cmd.Root().GenBashCompletion(cmd.OutOrStdout())
-			case "zsh":
-				return cmd.Root().GenZshCompletion(cmd.OutOrStdout())
-			case "fish":
-				return cmd.Root().GenFishCompletion(cmd.OutOrStdout(), true)
-			case "powershell":
-				return cmd.Root().GenPowerShellCompletionWithDesc(cmd.OutOrStdout())
-			default:
-				return fmt.Errorf("unknown shell: %s", args[0])
-			}
-		},
-	}
-}
